backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.2
googlemaps
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==4.2.27
django-sslserver==0.22
flake8==7.3.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
mccabe==0.7.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
reportlab==4.4.7
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.3.0
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.15.0
whitenoise==6.11.0
zope.interface==8.0.1
djangorestframework
